==> 2-coachella_process_pmf.Rmd <==
---
title: "coachella_process"
author: "ricardo_piedrahita"
knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_file = paste0(substr(inputFile,1,nchar(inputFile)-4),Sys.Date(),'.html'))})
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document:
    toc: true
    toc_float: true
    theme: united
    self_contained: yes
    code_folding: hide
---


# Script 2 of 3 for Coachella data analysis
# Load


```{r setup, include=TRUE,echo=TRUE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE, fig.width = 8, fig.height = 4)

bdlthresh = 60 # percent

graphics.off()

## Analyze Coachella ASPEN data
#Keys: instrument_id, location, size

source('r_scripts/load.R')
source('r_scripts/ASPEN_functions.R')
email = 0 #Set to 1 to send out summary qaqc email, else 0


```

# Import data
## SampleDate is the file_start_date
```{r import_processed, include=TRUE,echo=TRUE}

dataGrav = read_xlsx("~/Dropbox/Coachella Source Apportionment PHI/Data/FilterResultsCombined_MR.xlsx",
                     sheet = "Weights_Results",skip = 1) %>% 
  rename_with(., ~ gsub(" ","",.x)) %>% 
  rename_with(., ~ gsub("[[:punct:]]","",.x)) %>% 
  rename("samplevolm3"="Correctedsamplevolumeweightaccountsformissingfiles") %>% 
  dplyr::select(SampleDate,FilterSizeType,SampleType,SampleDurationweight,samplevolm3,
                Gravugm3,Gravuncugm3) %>% 
  dplyr::filter(!is.na(Gravugm3))

dataECOC = read_xlsx("~/Dropbox/Coachella Source Apportionment PHI/Data/FilterResultsCombined_MR.xlsx",
                     sheet = "ECOC_Results",skip = 5) %>% 
  rename_with(., ~ gsub(" ","",.x)) %>% 
  rename_with(., ~ gsub("[[:punct:]]","",.x)) %>% 
  dplyr::select(SampleDate,FilterSizeType,
                OCconcugm3,OCuncugm3,ECconcugm3,ECuncugm3) %>% 
  dplyr::filter(!is.na(SampleDate))


dataOrganics = read_xlsx("~/Dropbox/Coachella Source Apportionment PHI/Data/FilterResultsCombined_MR.xlsx",
                         sheet = "Organics_Results",skip = 3) %>% 
  rename_with(., ~ gsub(" ","",.x)) %>% 
  rename_with(., ~ gsub("[[:punct:]]","",.x)) %>% 
  dplyr::select(SampleDate,FilterSizeType,ends_with("conc"),ends_with("unc")) %>% 
  dplyr::filter(!is.na(SampleDate))


#Remove data with bdl data greater than 60%
dataXRF_bdlpercentage = read_xlsx("~/Dropbox/Coachella Source Apportionment PHI/Data/FilterResultsCombined_MR.xlsx",
                    sheet = "XRF_Results",skip = 0) %>% 
  rename_with(., ~ gsub(" ","",.x)) %>% 
  rename_with(., ~ gsub("/","",.x)) %>%
  rename_with(., ~ gsub("_","",.x)) %>%
  dplyr::select(ends_with("ugm3"),-starts_with("...")) %>% 
  dplyr::filter(row_number()==1) %>% 
  pivot_longer(cols = 1:22) %>% 
  dplyr::mutate(value = ifelse(value>bdlthresh/100,NA,value)) %>% 
  dplyr::filter(!is.na(value)) %>% 
  dplyr::mutate(name2 = gsub("ugm3","uncugm3",name))


dataXRF = read_xlsx("~/Dropbox/Coachella Source Apportionment PHI/Data/FilterResultsCombined_MR.xlsx",
                    sheet = "XRF_Results",skip = 4,) %>% 
  rename_with(., ~ gsub(" ","",.x)) %>% 
  rename_with(., ~ gsub("/","",.x)) %>%
  rename_with(., ~ gsub("_","",.x)) %>%
  dplyr::select(-starts_with("ugcm"),-"Correctedsamplevolumexrf(accountsformissingfiles)",-"SampleVolumexrf(m^3)",
                -"Blanksubtractedmassxrf(µg)",-"Sampledurationxrf(hours)",-"Guesses,mustupdate!",
                -FilterType,-FilterID,-SampleType,-starts_with("..."),
                -starts_with("Error"),-ends_with("ugcm2")) %>% 
  dplyr::filter(!is.na(SampleDate)) %>% 
  dplyr::mutate(FilterSizeType = as.factor(FilterSizeType)) %>% 
  dplyr::mutate_each(funs(if(is.character(.)) as.numeric(.) else .)) %>% 
  dplyr::select(starts_with("Sample"),starts_with("Filter"),
                starts_with(dput(as.character(dataXRF_bdlpercentage$name))),
                starts_with(dput(as.character(dataXRF_bdlpercentage$name2))))



dataAll = dataGrav  %>% 
  full_join(dataECOC, by = c("SampleDate","FilterSizeType")) %>% 
  left_join(dataXRF, by = c("SampleDate","FilterSizeType"))

```


# Plot up processed data
```{r plots_processed, include=TRUE,echo=TRUE}

dataLong = dataAll %>% 
  pivot_longer(cols = c(Gravugm3:Pbuncugm3)) %>% 
  dplyr::filter(!is.na(value)) %>% 
  dplyr::distinct()


give.n <- function(x){return(c(y = 0, label = length(x)))}
give.meanlog <- function(x){return(c(y =mean(x)+.2, label = round(10^(mean(x)),digits=3)))}

#To make box and whiskers quantiles rather than IQRs.
f <- function(x) {
  r <- quantile(x, probs = c(0.05, 0.25, 0.5, 0.75, 0.95))
  names(r) <- c("ymin", "lower", "middle", "upper", "ymax")
  r
}

distributions25 <- dataLong %>% 
  dplyr::filter(!name %like% "unc") %>% 
  dplyr::filter(!FilterSizeType %like% "10") %>% 
  ggplot(aes(x=name,y=value)) + 
  stat_summary(fun.data = f, geom="boxplot") +  
  stat_summary(fun=mean, colour="blue", geom="point", 
               shape=18, size=3,alpha = 0.8) +
  stat_summary(fun.data = give.meanlog, geom = "text",colour="blue",size=4) +
  geom_jitter(height = 0,width = 0.4,alpha = 0.05) +
  stat_summary(fun.data = give.n, geom = "text") +
  facet_grid(FilterSizeType ~ .) +
  labs(y=expression("µg/m^3"),x="") + 
  ggtitle("PM2.5 concentration distributions") + 
  # theme_minimal() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1)) +
  theme(legend.title = element_blank()) +
  scale_y_log10() 
distributions25

distributions10 <- dataLong %>% 
  dplyr::filter(!name %like% "unc") %>% 
  dplyr::filter(FilterSizeType %like% "10") %>% 
  ggplot(aes(x=name,y=value))+ 
  stat_summary(fun.data = f, geom="boxplot") +  
  stat_summary(fun=mean, colour="blue", geom="point", 
               shape=18, size=3,alpha = 0.8) +
  stat_summary(fun.data = give.meanlog, geom = "text",colour="blue",size=4) +
  geom_jitter(height = 0,width = 0.4,alpha = 0.05) +
  stat_summary(fun.data = give.n, geom = "text") +
  facet_grid(FilterSizeType ~ .) +
  labs(y=expression("µg/m^3"),x="") + 
  ggtitle("PM10 concentration distributions") + 
  # theme_minimal() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1)) +
  theme(legend.title = element_blank()) +
  scale_y_log10() 
distributions10

tseries25 <- dataLong %>% 
  dplyr::filter(!name %like% "unc") %>% 
  dplyr::filter(!FilterSizeType %like% "10") %>% 
  ggplot(aes(x=SampleDate,y=value))+ 
  geom_smooth(alpha = 0.25) + 
  geom_point(alpha = 0.25) +
  facet_wrap(name ~ .,scales="free_y") +
  labs(y=expression("µg/m^3"),x="") + 
  ggtitle("Indio PM2.5 species concentration time series") + 
  # theme_minimal() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1)) +
  theme(legend.title = element_blank())
tseries25


tseries10 <- dataLong %>% 
  dplyr::filter(!name %like% "unc") %>% 
  dplyr::filter(FilterSizeType %like% "10") %>% 
  ggplot(aes(x=SampleDate,y=value))+ 
  geom_smooth(alpha = 0.25) + 
  geom_point(alpha = 0.25) +
  facet_wrap(name ~ .,scales="free_y") +
  labs(y=expression("µg/m^3"),x="") + 
  ggtitle("Indio PM10 species concentration time series") + 
  # theme_minimal() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1)) +
  theme(legend.title = element_blank())
tseries10

```



# Save processed data in PMF format
```{r save_processed, include=TRUE,echo=TRUE}


dataConcWide25 = dataLong %>% 
  dplyr::filter(!name %like% "unc") %>% 
  dplyr::distinct()   %>% 
  dplyr::filter(!FilterSizeType %like% "10") %>% 
  pivot_wider(values_from = value,names_from = c("name")) %>% 
  dplyr::select(-c("FilterSizeType","SampleType","SampleDurationweight","samplevolm3")) %>% 
  rename(setNames(names(.), tolower(gsub("ugm3", "", names(.))))) %>% 
  rename(setNames(names(.), tolower(gsub("conc", "", names(.)))))%>% 
  dplyr::arrange(sampledate)


write.csv(dataConcWide25,paste0("pmf_formatted_data/conc_pm25_",Sys.Date(),".csv"), row.names = FALSE)
write.csv(dataConcWide25[complete.cases(dataConcWide25),],paste0("pmf_formatted_data/conc_pm25_completecases_",Sys.Date(),".csv"), row.names = FALSE)


dataConcWide10 = dataLong %>% 
  dplyr::filter(!name %like% "unc") %>% 
  dplyr::distinct()   %>% 
  dplyr::filter(FilterSizeType %like% "10") %>% 
  pivot_wider(values_from = value) %>% 
  dplyr::select(-c("FilterSizeType","SampleType","SampleDurationweight","samplevolm3")) %>% 
  rename(setNames(names(.), tolower(gsub("ugm3", "", names(.))))) %>% 
  rename(setNames(names(.), tolower(gsub("unc", "", names(.))))) %>% 
  rename(setNames(names(.), tolower(gsub("conc", "", names(.)))))%>% 
  dplyr::arrange(sampledate)
write.csv(dataConcWide10,paste0("pmf_formatted_data/conc_pm10_",Sys.Date(),".csv"), row.names = FALSE)
write.csv(dataConcWide10[complete.cases(dataConcWide10),],paste0("pmf_formatted_data/conc_pm10_completecases_",Sys.Date(),".csv"), row.names = FALSE)


dataUncWide25 = dataLong %>% 
  dplyr::filter(name %like% "unc") %>% 
  dplyr::distinct()   %>% 
  dplyr::filter(!FilterSizeType %like% "10") %>% 
  pivot_wider(values_from = value) %>% 
  dplyr::select(-c("FilterSizeType","SampleType","SampleDurationweight","samplevolm3")) %>% 
  rename(setNames(names(.), tolower(gsub("ugm3", "", names(.))))) %>% 
  rename(setNames(names(.), tolower(gsub("unc", "", names(.))))) %>% 
  rename(setNames(names(.), tolower(gsub("conc", "", names(.))))) %>% 
  dplyr::arrange(sampledate)
write.csv(dataUncWide25,paste0("pmf_formatted_data/unc_pm25_",Sys.Date(),".csv"), row.names = FALSE)
write.csv(dataUncWide25[complete.cases(dataUncWide25),],paste0("pmf_formatted_data/unc_pm25_completecases_",Sys.Date(),".csv"), row.names = FALSE)

dataUncWide10 = dataLong %>% 
  dplyr::filter(name %like% "unc") %>% 
  dplyr::filter(FilterSizeType %like% "10") %>% 
  pivot_wider(values_from = value) %>% 
  dplyr::select(-c("FilterSizeType","SampleType","SampleDurationweight","samplevolm3")) %>% 
  rename(setNames(names(.), tolower(gsub("ugm3", "", names(.))))) %>% 
  rename(setNames(names(.), tolower(gsub("unc", "", names(.))))) %>% 
  rename(setNames(names(.), tolower(gsub("conc", "", names(.)))))%>% 
  dplyr::arrange(sampledate)
write.csv(dataUncWide10,paste0("pmf_formatted_data/unc_pm10_",Sys.Date(),".csv"), row.names = FALSE)
write.csv(dataUncWide10[complete.cases(dataUncWide10),],paste0("pmf_formatted_data/unc_pm10_completecases_",Sys.Date(),".csv"), row.names = FALSE)


```
